refactor: Order update 로직을 Trade 도메인에서 처리하도록 변경
